Reduce overhead:
The function scgpt_utils.get_embeddings is
completely refactored:
The processing of the data (part of which was also
in tasks.cell_emb.get_batch_cell_embeddings) is
now in Helicals model.process_data function. It
now returns the Dataset object. Which is then an
input to the model.get_embeddings function.
For now, I completely removed the option whether
to return a new AnnData object after getting the
embeddings where the cell embeddings are added to
a new :attr:`adata.obsm` with key "X_scGPT".